@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node ./public/server.js",
-    "build": "webpack",
+    "build": "webpack --mode development",
     "dist": "webpack",
     "start-dev": "NODE_ENV=development npm run build && npm run start",
     "serve": "NODE_ENV=development webpack-dev-server --hot"
@@ -15,8 +15,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "axios": "^1.3.5",
     "babel-eslint": "^10.1.0",
     "express": "^4.17.1",
+    "node-fetch": "^3.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-side-effect": "^2.1.2",
@@ -50,10 +52,12 @@
     "node-sass": "^7.0.1",
     "prettier": "^2.8.7",
     "react-test-renderer": "^17.0.2",
-    "sass-loader": "^13.0.2",
+    "sass": "^1.61.0",
+    "sass-loader": "^13.2.2",
+    "style-loader": "^3.3.2",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
-    "webpack": "^5.37.1",
+    "webpack": "^5.78.0",
     "webpack-cli": "^4.7.0",
     "webpack-dev-server": "^4.13.2",
     "webpack-merge": "^5.8.0",
