@@ -1,7 +1,21 @@
-const React = require('react');
-const About = () => (
-  <h1>About</h1>
-);
+import React, { useState } from 'react';
+
+const Home = ({name}) => {
+  const [changeValue, setChangeValue] = useState('ON')
+  const toogle = () => setChangeValue((prev) => prev === 'ON' ? 'OFF' : 'ON')
+  return (
+    <>
+      <h1>Home2 {name} {changeValue}</h1>
+      <button onClick={toogle}>Toogle</button>
+    </>
+  );
+};
+
+export const getServerSideProps = async (req, res) => {
+  return {
+    name: 'Hernan Correa - CO'
+  };
+};
 
 // export default About;
-module.exports = About;
+export default Home;
