@@ -6,14 +6,14 @@ const Page = ({ state, module, children }) => {
   const serializedProps = JSON.stringify(state).replace(/</g, '\\u003c');
   return (
     <>
-      <LinkTag href={`/static/${module}.bundle.css`} />
+      <LinkTag href={`/${module}.css.bundle.css`} />
       <ScriptTag>
         {`
           window.__INITIAL_PROPS__ = ${serializedProps};
         `}
       </ScriptTag>
-      <ScriptTag src={`/static/${module}.bundle.js`} async defer/>
-      <ScriptTag src={`/static/vendor.bundle.js`} async defer/>
+      <ScriptTag src={`/${module}.bundle.js`} async defer/>
+      <ScriptTag src={`/vendor.bundle.js`} async defer/>
       {children}
     </>
   )
